LayoutTests/accessibility/contenteditable-hidden-div.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
</head>
<body id="body">

<div contentEditable=true style="outline:1px solid"></div>

<h3 id="h3" tabindex=0>test</h3>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that a contenteditable element will not be ignored by accessibility.");

    if (window.accessibilityController) {

        document.getElementById("body").focus();

        // make sure this element is NOT the h3 element (which otherwise would be the first element)
        var editableDiv = accessibilityController.focusedElement.childAtIndex(0);

        document.getElementById("h3").focus();
        shouldBe("accessibilityController.focusedElement.isEqual(editableDiv)", "false");
    }

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
