LayoutTests/platform/blackberry/accessibility/contenteditable-hidden-div-expected.txt

test

This tests that a contenteditable element will not be ignored by accessibility.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL accessibilityController.focusedElement.isEqual(editableDiv) should be false. Was true.
PASS successfullyParsed is true

TEST COMPLETE
